Fix integer overflow in toMB() function.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@983 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/utils.h
@@ -211,7 +211,7 @@ void create_directory(const std::string& path);
 // Modified julian day
 int mjd(int day, int month, int year);
 
-inline double toMB(unsigned int size) {
+inline double toMB(unsigned long long size) {
   return (1.0*size)/(1024*1024);
 }
